feat(subscribe): make the watcher 'pre' by default like regular watchers
`store.$subscribe()` now takes an extra argument like `watch` and
provides a more flexible API. You can now set `deep`, `flush`, and other
`watch` options.

Close #610

BREAKING CHANGE: `$store.subscribe()` now runs a `flush: 'pre'` instead
of a `flush: 'sync'` by default. This is due to the performance
implications of using `deep: true` + `flush: 'sync'`. **Most of the
times, this change should not affect you**. If you were relying on the
`flush` being `sync`, you can still provide it as a second argument
which is now an object instead of a boolean (`detached` is now a
property of that object). Migrating this should be fairly easy as the
boolean version will show a warning in dev mode and TypeScript will show
the function as deprecated.

```diff
-store.$subscribe(callback, true)
+store.$subscribe(callback, { detached: true })

 // to keep the flush: 'sync' behavior
-store.$subscribe(callback)
+store.$subscribe(callback, { flush: 'sync' })
```